requirements.txt
discord.py
groq
flask
python-dotenv